+++ src/game/lib/components/Meter.svelte
