+++ src/game/lib/components/NewGame.svelte
