+++ src/game/lib/components/Spinner.svelte
