+++ src/game/lib/components/Wheel.svelte
