+++ src/game/lib/constants.ts
