+++ src/game/lib/images/svelte-welcome.png
