+++ src/game/lib/images/svelte-welcome.webp
