+++ src/game/lib/index.ts
