+++ src/game/main.ts
@@ -4,28 +4,23 @@ import { Game as MainGame } from './scenes/Game';
 import { MainMenu } from './scenes/MainMenu';
 import { AUTO, Game } from 'phaser';
 import { Preloader } from './scenes/Preloader';
+import { initFlowbite } from 'flowbite';
+
+import { gameParams, gameState, loading } from '../stores.ts';
 
 //  Find out more information about the Game Config at:
 //  https://newdocs.phaser.io/docs/3.70.0/Phaser.Types.Core.GameConfig
 const config: Phaser.Types.Core.GameConfig = {
-    type: AUTO,
-    width: 1024,
-    height: 768,
-    parent: 'game-container',
-    backgroundColor: '#028af8',
-    scene: [
-        Boot,
-        Preloader,
-        MainMenu,
-        MainGame,
-        GameOver
-    ]
+	type: AUTO,
+	width: 1024,
+	height: 768,
+	parent: 'game-container',
+	backgroundColor: '#028af8',
+	scene: [Boot, Preloader, MainMenu, MainGame, GameOver]
 };
 
 const StartGame = (parent: string) => {
-
-    return new Game({ ...config, parent });
-
-}
+	return new Game({ ...config, parent });
+};
 
 export default StartGame;
