+++ src/game/scenes/Preloader.ts
@@ -1,47 +1,61 @@
 import { Scene } from 'phaser';
+import { SVG } from '@svgdotjs/svg.js';
+import houseSvg from '$lib/images/house.svg?raw';
+import buildingSvg from '$lib/images/building.svg?raw';
+import buildingDomeSvg from '$lib/images/building-dome.svg?raw';
+import walkSvg from '$lib/images/walk.svg?raw';
+import citySvg from '$lib/images/city-map.svg?raw';
+import parkSvg from '$lib/images/park.svg?raw';
+import discoSvg from '$lib/images/disco.svg?raw';
+import personSvg from '$lib/images/person.svg?raw';
+import Menu from '$lib/components/Menu.svelte';
+import { browser } from '$app/environment';
+import { goto } from '$app/navigation';
+import { Button, Modal } from 'flowbite-svelte';
+import { onMount, onDestroy } from 'svelte';
+import { MS_PER_HOUR } from '../lib/constants.ts';
+import { engine } from '../engine/engine.ts';
 
-export class Preloader extends Scene
-{
-    constructor ()
-    {
-        super('Preloader');
-    }
+// some component level global variables
+const buildingPositions = $gameParams.defaults.buildingPositions;
 
-    init ()
-    {
-        //  We loaded this image in our Boot Scene, so we can display it here
-        this.add.image(512, 384, 'background');
+export class Preloader extends Scene {
+	constructor() {
+		super('Preloader');
+	}
 
-        //  A simple progress bar. This is the outline of the bar.
-        this.add.rectangle(512, 384, 468, 32).setStrokeStyle(1, 0xffffff);
+	init() {
+		//  We loaded this image in our Boot Scene, so we can display it here
+		this.add.image(512, 384, 'background');
 
-        //  This is the progress bar itself. It will increase in size from the left based on the % of progress.
-        const bar = this.add.rectangle(512-230, 384, 4, 28, 0xffffff);
+		//  A simple progress bar. This is the outline of the bar.
+		this.add.rectangle(512, 384, 468, 32).setStrokeStyle(1, 0xffffff);
 
-        //  Use the 'progress' event emitted by the LoaderPlugin to update the loading bar
-        this.load.on('progress', (progress: number) => {
+		//  This is the progress bar itself. It will increase in size from the left based on the % of progress.
+		const bar = this.add.rectangle(512 - 230, 384, 4, 28, 0xffffff);
 
-            //  Update the progress bar (our bar is 464px wide, so 100% = 464px)
-            bar.width = 4 + (460 * progress);
+		//  Use the 'progress' event emitted by the LoaderPlugin to update the loading bar
+		this.load.on('progress', (progress: number) => {
+			//  Update the progress bar (our bar is 464px wide, so 100% = 464px)
+			bar.width = 4 + 460 * progress;
+		});
+	}
 
-        });
-    }
+	preload() {
+		//  Load the assets for the game - Replace with your own assets
+		this.load.setPath('assets');
 
-    preload ()
-    {
-        //  Load the assets for the game - Replace with your own assets
-        this.load.setPath('assets');
+		this.load.image('logo', 'logo.png');
+		this.load.image('star', 'star.png');
+	}
 
-        this.load.image('logo', 'logo.png');
-        this.load.image('star', 'star.png');
-    }
+	create() {
+		//  When all the assets have loaded, it's often worth creating global objects here that the rest of the game can use.
+		//  For example, you can define global animations here, so we can use them in other scenes.
 
-    create ()
-    {
-        //  When all the assets have loaded, it's often worth creating global objects here that the rest of the game can use.
-        //  For example, you can define global animations here, so we can use them in other scenes.
-
-        //  Move to the MainMenu. You could also swap this for a Scene Transition, such as a camera fade.
-        this.scene.start('MainMenu');
-    }
+		//  Move to the MainMenu. You could also swap this for a Scene Transition, such as a camera fade.
+		this.scene.start('MainMenu');
+	}
 }
+
+export default Preloader;
